Move GetClusterProxyConn in proxy.go was in cluster.go Add variable cluster.haveDBTLScert MySQL/MariaDB connection in monitor to use TLS
